fix and refactor combineFontStyleAndFontWeight
 - update complete-organization-chart-new to use a basic font
